requirements.txt
fastapi
google-ai-generativelanguage==0.6.2
langchain==0.1.17
pillow
retry==0.9.2
uvicorn
urlextract==1.9.0